# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ActorState": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "orientation_deg": {
            "items": {
              "type": "number"
            },
            "maxItems": 3,
            "minItems": 3,
            "title": "Orientation Deg",
            "type": "array"
          },
          "position": {
            "items": {
              "type": "number"
            },
            "maxItems": 3,
            "minItems": 3,
            "title": "Position",
            "type": "array"
          }
        },
        "required": [
          "id",
          "position"
        ],
        "title": "ActorState",
        "type": "object"
      },
...
      "DeviceState": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "position": {
            "items": {
              "type": "number"
            },
            "maxItems": 3,
            "minItems": 3,
            "title": "Position",
            "type": "array"
          }
        },
        "required": [
          "id",
          "position"
        ],
        "title": "DeviceState",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LinkMetrics": {
        "additionalProperties": false,
        "properties": {
          "interference_dbm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interference Dbm"
          },
          "path_count": {
            "default": 0,
            "title": "Path Count",
            "type": "integer"
          },
          "path_gain_db": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path Gain Db"
          },
          "rms_delay_spread_ns": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rms Delay Spread Ns"
          },
          "rss_dbm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rss Dbm"
          },
          "rx_id": {
            "title": "Rx Id",
            "type": "string"
          },
          "sinr_db": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sinr Db"
          },
          "tx_id": {
            "title": "Tx Id",
            "type": "string"
          }
        },
        "required": [
          "tx_id",
          "rx_id"
        ],
        "title": "LinkMetrics",
        "type": "object"
      },
...
      "PathInteraction": {
        "additionalProperties": false,
        "properties": {
          "point": {
            "items": {
              "type": "number"
            },
            "maxItems": 3,
            "minItems": 3,
            "title": "Point",
            "type": "array"
          },
          "prim_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prim Id"
          },
          "rf_material_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rf Material Id"
          },
          "type": {
            "enum": [
              "reflection",
              "diffraction",
              "scattering",
              "transmission"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "point"
        ],
        "title": "PathInteraction",
        "type": "object"
      },
...
      "RadioMapGridConfig": {
        "additionalProperties": false,
        "properties": {
          "cell_size_m": {
            "default": 2.0,
            "exclusiveMinimum": 0.0,
            "title": "Cell Size M",
            "type": "number"
          },
          "center_xy": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "maxItems": 2,
                "minItems": 2,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Center Xy"
          },
          "height_m": {
            "default": 1.5,
            "title": "Height M",
            "type": "number"
          },
          "metric": {
            "default": "path_gain_db",
            "enum": [
              "path_gain_db",
              "rss_dbm",
              "sinr_db"
            ],
            "title": "Metric",
            "type": "string"
          },
          "size_xy": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "maxItems": 2,
                "minItems": 2,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Xy"
          }
        },
        "title": "RadioMapGridConfig",
        "type": "object"
      },
...
      "RayPath": {
        "additionalProperties": false,
        "properties": {
          "aoa_deg": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aoa Deg"
          },
          "aod_deg": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aod Deg"
          },
          "delay_ns": {
            "minimum": 0.0,
            "title": "Delay Ns",
            "type": "number"
          },
          "interactions": {
            "items": {
              "$ref": "#/components/schemas/PathInteraction"
            },
            "title": "Interactions",
            "type": "array"
          },
          "path_gain_db": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path Gain Db"
          },
          "path_id": {
            "title": "Path Id",
            "type": "string"
          },
          "path_type": {
            "enum": [
              "los",
              "reflection",
              "diffraction",
              "scattering",
              "transmission",
              "mixed"
            ],
            "title": "Path Type",
            "type": "string"
          },
          "phase_rad": {
            "default": 0.0,
            "title": "Phase Rad",
            "type": "number"
          },
          "power_dbm": {
            "title": "Power Dbm",
            "type": "number"
          },
          "rx_id": {
            "title": "Rx Id",
            "type": "string"
          },
          "tx_id": {
            "title": "Tx Id",
            "type": "string"
          },
          "vertices": {
            "items": {
              "items": {
                "type": "number"
              },
              "maxItems": 3,
              "minItems": 3,
              "type": "array"
            },
            "minItems": 2,
            "title": "Vertices",
            "type": "array"
          }
        },
        "required": [
          "path_id",
          "tx_id",
          "rx_id",
          "path_type",
          "vertices",
          "power_dbm",
          "delay_ns"
        ],
        "title": "RayPath",
        "type": "object"
      },
...
      "ScenarioFrame": {
        "additionalProperties": false,
        "properties": {
          "actor_states": {
            "items": {
              "$ref": "#/components/schemas/ActorState"
            },
            "title": "Actor States",
            "type": "array"
          },
          "device_states": {
            "items": {
              "$ref": "#/components/schemas/DeviceState"
            },
            "title": "Device States",
            "type": "array"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/LinkMetrics"
            },
            "title": "Links",
            "type": "array"
          },
          "paths": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RayPath"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paths"
          },
          "time_s": {
            "title": "Time S",
            "type": "number"
          }
        },
        "required": [
          "time_s"
        ],
        "title": "ScenarioFrame",
        "type": "object"
      },
      "ScenarioResultSet": {
        "additionalProperties": false,
        "properties": {
          "backend": {
            "title": "Backend",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/ScenarioFrame"
            },
            "title": "Frames",
            "type": "array"
          },
          "kind": {
            "const": "scenario",
            "default": "scenario",
            "title": "Kind",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "result_id": {
            "title": "Result Id",
            "type": "string"
          },
          "simulation_config_id": {
            "title": "Simulation Config Id",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "result_id",
          "backend",
          "simulation_config_id"
        ],
        "title": "ScenarioResultSet",
        "type": "object"
      },
      "ScenarioSimulateRequest": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimulationConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "config_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Id"
          },
          "dt_s": {
            "default": 0.1,
            "exclusiveMinimum": 0.0,
            "title": "Dt S",
            "type": "number"
          },
          "include_paths": {
            "default": true,
            "title": "Include Paths",
            "type": "boolean"
          },
          "num_frames": {
            "default": 20,
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Num Frames",
            "type": "integer"
          }
        },
        "title": "ScenarioSimulateRequest",
        "type": "object"
      },
...
      "SimulationConfig": {
        "additionalProperties": false,
        "description": "Full Sionna RT solver parameter surface (mirrors PathSolver/\nRadioMapSolver options so every knob is user-controllable in the UI).",
        "properties": {
          "backend": {
            "default": "auto",
            "enum": [
              "auto",
              "mock",
              "sionna"
            ],
            "title": "Backend",
            "type": "string"
          },
          "bandwidth_hz": {
            "default": 100000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Bandwidth Hz",
            "type": "number"
          },
          "diffraction": {
            "default": false,
            "title": "Diffraction",
            "type": "boolean"
          },
          "diffraction_lit_region": {
            "default": false,
            "title": "Diffraction Lit Region",
            "type": "boolean"
          },
          "edge_diffraction": {
            "default": false,
            "title": "Edge Diffraction",
            "type": "boolean"
          },
          "engine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine"
          },
          "frequency_hz": {
            "default": 28000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Frequency Hz",
            "type": "number"
          },
          "id": {
            "default": "default",
            "title": "Id",
            "type": "string"
          },
          "los": {
            "default": true,
            "title": "Los",
            "type": "boolean"
          },
          "max_depth": {
            "default": 3,
            "maximum": 12.0,
            "minimum": 0.0,
            "title": "Max Depth",
            "type": "integer"
          },
          "name": {
            "default": "Default",
            "title": "Name",
            "type": "string"
          },
          "noise_figure_db": {
            "default": 7.0,
            "minimum": 0.0,
            "title": "Noise Figure Db",
            "type": "number"
          },
          "num_samples": {
            "default": 1000000,
            "minimum": 1.0,
            "title": "Num Samples",
            "type": "integer"
          },
          "radio_map": {
            "$ref": "#/components/schemas/RadioMapGridConfig"
          },
          "reflection": {
            "default": true,
            "title": "Reflection",
            "type": "boolean"
          },
          "refraction": {
            "default": false,
            "title": "Refraction",
            "type": "boolean"
          },
          "rx_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rx Ids"
          },
          "scattering": {
            "default": false,
            "title": "Scattering",
            "type": "boolean"
          },
          "seed": {
            "default": 42,
            "minimum": 0.0,
            "title": "Seed",
            "type": "integer"
          },
          "synthetic_array": {
            "default": true,
            "title": "Synthetic Array",
            "type": "boolean"
          },
          "tx_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Ids"
          }
        },
        "title": "SimulationConfig",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Unified RF-visual scene authoring, RF material assignment, Sionna RT projection compilation, and simulation result APIs.",
    "title": "SionnaTwin Studio",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects/{project_id}/results/scenario": {
      "get": {
        "operationId": "get_scenario_result_api_projects__project_id__results_scenario_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "result_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResultSet"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scenario Result",
        "tags": [
          "scenario"
        ]
      }
    },
...
    "/api/projects/{project_id}/simulate/scenario": {
      "post": {
        "operationId": "simulate_scenario_api_projects__project_id__simulate_scenario_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ScenarioSimulateRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResultSet"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Scenario",
        "tags": [
          "scenario"
        ]
      }
    },
...
  }
}